+++ lib/arel/extensions.rb
@@ -12,6 +12,7 @@ require 'arel/extensions/user'
 require 'arel/extensions/current_catalog'
 require 'arel/extensions/current_schema'
 require 'arel/extensions/array'
+require 'arel/extensions/indirection'
 
 require 'arel/extensions/functions'
 require 'arel/extensions/nodes'
